modif

+++ index.php
@@ -231,17 +231,17 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             //if (isset($result['status']) && $result['status'] === 'erreur'){
                 flash('error', 'Partie non supprimée.');
                 redirectTo('admin');
-          /*  }
-            else */{
+            //}
+            if(true) {
                 $stmt = $pdo->prepare("DELETE FROM parties WHERE id = ?");
                 $stmt->execute([(int)($_POST['id'] ?? 0)]);
-                /*foreach($result as $player){
+                foreach($result as $player){
                     $pseudo = $player[0];
                     $score = intval($player[1]);
 
                     $scrpt = $pdo->prepare("UPDATE joueurs SET score = score + ?, kills = kills + ?, matchs = matchs + 1 WHERE pseudo = ? ");
                     $scrpt->execute([$score,$score,$pseudo]);
-                }*/
+                }
                 flash('success', 'Partie supprimée.');
                 redirectTo('admin');
             }
